requirements.txt
CherryPy==3.8.0
Flask==0.10.1
Flask-RESTful==0.3.4
parsedatetime==1.5
Paste==2.0.2
pyslack==0.1.3
requests==2.7.0
